# openapi.yaml
openapi: 3.1.0
info:
  title: Qarax API
  description: REST API for managing virtual machines and hypervisor hosts
  license:
    name: ''
  version: 0.1.0
paths:
  /boot-sources:
    get:
      tags:
      - boot-sources
      operationId: list
      responses:
        '200':
          description: List all boot sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BootSource'
        '500':
          description: Internal server error
    post:
      tags:
      - boot-sources
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewBootSource'
        required: true
      responses:
        '201':
          description: Boot source created successfully
          content:
            text/plain:
              schema:
                type: string
        '422':
          description: Invalid input
        '500':
          description: Internal server error
  /boot-sources/{boot_source_id}:
    get:
      tags:
      - boot-sources
      operationId: get
      parameters:
      - name: boot_source_id
        in: path
        description: Boot source unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Boot source found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootSource'
        '404':
          description: Boot source not found
        '500':
          description: Internal server error
    delete:
      tags:
      - boot-sources
      operationId: delete
      parameters:
      - name: boot_source_id
        in: path
        description: Boot source unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Boot source deleted successfully
        '404':
          description: Boot source not found
        '500':
          description: Internal server error
  /hosts:
    get:
      tags:
      - hosts
      operationId: list
      responses:
        '200':
          description: List all hosts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Host'
        '500':
          description: Internal server error
    post:
      tags:
      - hosts
      operationId: add
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewHost'
        required: true
      responses:
        '201':
          description: Host created successfully
          content:
            text/plain:
              schema:
                type: string
        '409':
          description: Host with name already exists
        '422':
          description: Invalid input
        '500':
          description: Internal server error
  /hosts/{host_id}:
    patch:
      tags:
      - hosts
      operationId: update
      parameters:
      - name: host_id
        in: path
        description: Host unique identifier
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateHostRequest'
        required: true
      responses:
        '200':
          description: Host updated successfully
        '404':
          description: Host not found
        '500':
          description: Internal server error
  /hosts/{host_id}/deploy:
    post:
      tags:
      - hosts
      operationId: deploy
      parameters:
      - name: host_id
        in: path
        description: Host unique identifier
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeployHostRequest'
        required: true
      responses:
        '202':
          description: Host deployment accepted
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Host not found
        '422':
          description: Invalid input
        '500':
          description: Internal server error
  /hosts/{host_id}/init:
    post:
      tags:
      - hosts
      operationId: init
      parameters:
      - name: host_id
        in: path
        description: Host unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Host initialized successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Host'
        '404':
          description: Host not found
        '500':
          description: Internal server error
  /jobs/{job_id}:
    get:
      tags:
      - jobs
      operationId: get
      parameters:
      - name: job_id
        in: path
        description: Job unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          description: Job not found
        '500':
          description: Internal server error
  /storage-objects:
    get:
      tags:
      - storage-objects
      operationId: list
      responses:
        '200':
          description: List all storage objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StorageObject'
        '500':
          description: Internal server error
    post:
      tags:
      - storage-objects
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewStorageObject'
        required: true
      responses:
        '201':
          description: Storage object created successfully
          content:
            text/plain:
              schema:
                type: string
        '422':
          description: Invalid input
        '500':
          description: Internal server error
  /storage-objects/{object_id}:
    get:
      tags:
      - storage-objects
      operationId: get
      parameters:
      - name: object_id
        in: path
        description: Storage object unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Storage object found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageObject'
        '404':
          description: Storage object not found
        '500':
          description: Internal server error
    delete:
      tags:
      - storage-objects
      operationId: delete
      parameters:
      - name: object_id
        in: path
        description: Storage object unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Storage object deleted successfully
        '404':
          description: Storage object not found
        '500':
          description: Internal server error
  /storage-pools:
    get:
      tags:
      - storage-pools
      operationId: list
      responses:
        '200':
          description: List all storage pools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StoragePool'
        '500':
          description: Internal server error
    post:
      tags:
      - storage-pools
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewStoragePool'
        required: true
      responses:
        '201':
          description: Storage pool created successfully
          content:
            text/plain:
              schema:
                type: string
        '422':
          description: Invalid input
        '500':
          description: Internal server error
  /storage-pools/{pool_id}:
    get:
      tags:
      - storage-pools
      operationId: get
      parameters:
      - name: pool_id
        in: path
        description: Storage pool unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Storage pool found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoragePool'
        '404':
          description: Storage pool not found
        '500':
          description: Internal server error
    delete:
      tags:
      - storage-pools
      operationId: delete
      parameters:
      - name: pool_id
        in: path
        description: Storage pool unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Storage pool deleted successfully
        '404':
          description: Storage pool not found
        '500':
          description: Internal server error
  /storage-pools/{pool_id}/hosts:
    post:
      tags:
      - storage-pools
      operationId: attach_host
      parameters:
      - name: pool_id
        in: path
        description: Storage pool unique identifier
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachHostRequest'
        required: true
      responses:
        '204':
          description: Host attached to storage pool
        '404':
          description: Storage pool or host not found
        '500':
          description: Internal server error
  /storage-pools/{pool_id}/hosts/{host_id}:
    delete:
      tags:
      - storage-pools
      operationId: detach_host
      parameters:
      - name: pool_id
        in: path
        description: Storage pool unique identifier
        required: true
        schema:
          type: string
          format: uuid
      - name: host_id
        in: path
        description: Host unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Host detached from storage pool
        '404':
          description: Host not found
        '500':
          description: Internal server error
  /storage-pools/{pool_id}/import:
    post:
      tags:
      - storage-pools
      summary: Import an OCI image into the pool, converting it to OverlayBD format.
      operationId: import_to_pool
      parameters:
      - name: pool_id
        in: path
        description: Storage pool ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportToPoolRequest'
        required: true
      responses:
        '202':
          description: Import job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportToPoolResponse'
        '404':
          description: Pool not found
        '422':
          description: No UP host attached to pool
        '500':
          description: Internal server error
  /storage-pools/{pool_id}/transfers:
    get:
      tags:
      - transfers
      operationId: list
      parameters:
      - name: pool_id
        in: path
        description: Storage pool unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: List transfers for this pool
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transfer'
        '500':
          description: Internal server error
    post:
      tags:
      - transfers
      operationId: create
      parameters:
      - name: pool_id
        in: path
        description: Storage pool unique identifier
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTransfer'
        required: true
      responses:
        '202':
          description: Transfer accepted and started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '404':
          description: Storage pool not found
        '422':
          description: Invalid input
        '500':
          description: Internal server error
  /storage-pools/{pool_id}/transfers/{transfer_id}:
    get:
      tags:
      - transfers
      operationId: get
      parameters:
      - name: pool_id
        in: path
        description: Storage pool unique identifier
        required: true
        schema:
          type: string
          format: uuid
      - name: transfer_id
        in: path
        description: Transfer unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Transfer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '404':
          description: Transfer not found
        '500':
          description: Internal server error
  /vms:
    get:
      tags:
      - vms
      operationId: list
      responses:
        '200':
          description: List all VMs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vm'
        '500':
          description: Internal server error
    post:
      tags:
      - vms
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewVm'
        required: true
      responses:
        '201':
          description: VM created successfully (synchronous)
          content:
            application/json:
              schema:
                type: string
        '202':
          description: VM creation started asynchronously
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateVmResponse'
        '422':
          description: Invalid input
        '500':
          description: Internal server error
  /vms/{vm_id}:
    get:
      tags:
      - vms
      operationId: get
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: VM details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vm'
        '404':
          description: VM not found
        '500':
          description: Internal server error
    delete:
      tags:
      - vms
      operationId: delete
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: VM deleted successfully
        '404':
          description: VM not found
        '500':
          description: Internal server error
  /vms/{vm_id}/console:
    get:
      tags:
      - vms
      operationId: console_log
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Console log content
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Console logging not available for this VM
        '500':
          description: Internal server error
  /vms/{vm_id}/console/attach:
    get:
      tags:
      - vms
      summary: WebSocket console attachment for interactive terminal access
      operationId: console_attach
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '101':
          description: Switching Protocols - WebSocket connection established
        '404':
          description: VM not found
        '422':
          description: VM console not available or not in PTY mode
        '500':
          description: Internal server error
  /vms/{vm_id}/disks:
    post:
      tags:
      - vms
      summary: Attach a storage object to a VM as a disk.
      operationId: attach_disk
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttachDiskRequest'
        required: true
      responses:
        '201':
          description: Disk attached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VmDisk'
        '404':
          description: VM or storage object not found
        '422':
          description: VM not in Created state
        '500':
          description: Internal server error
...
  /vms/{vm_id}/pause:
    post:
      tags:
      - vms
      operationId: pause
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: VM paused successfully
        '404':
          description: VM not found
        '500':
          description: Internal server error
  /vms/{vm_id}/resume:
    post:
      tags:
      - vms
      operationId: resume
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: VM resumed successfully
        '404':
          description: VM not found
        '500':
          description: Internal server error
  /vms/{vm_id}/start:
    post:
      tags:
      - vms
      operationId: start
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '202':
          description: VM start accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VmStartResponse'
        '404':
          description: VM not found
        '422':
          description: VM not in a startable state
        '500':
          description: Internal server error
  /vms/{vm_id}/stop:
    post:
      tags:
      - vms
      operationId: stop
      parameters:
      - name: vm_id
        in: path
        description: VM unique identifier
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: VM stopped successfully
        '404':
          description: VM not found
        '500':
          description: Internal server error
components:
  schemas:
...
    BootSource:
      type: object
      required:
      - id
      - name
      - kernel_image_id
      properties:
        description:
          type:
          - string
          - 'null'
        id:
          type: string
          format: uuid
        initrd_image_id:
          type:
          - string
          - 'null'
          format: uuid
        kernel_image_id:
          type: string
          format: uuid
        kernel_params:
          type:
          - string
          - 'null'
        name:
          type: string
...
    Host:
      type: object
      required:
      - id
      - name
      - address
      - port
      - status
      - host_user
      properties:
        address:
          type: string
        available_memory_bytes:
          type:
          - integer
          - 'null'
          format: int64
        cloud_hypervisor_version:
          type:
          - string
          - 'null'
        disk_available_bytes:
          type:
          - integer
          - 'null'
          format: int64
        disk_total_bytes:
          type:
          - integer
          - 'null'
          format: int64
        host_user:
          type: string
        id:
          type: string
          format: uuid
        kernel_version:
          type:
          - string
          - 'null'
        load_average:
          type:
          - number
          - 'null'
          format: double
        name:
          type: string
        port:
          type: integer
          format: int32
        resources_updated_at:
          type:
          - string
          - 'null'
          format: date-time
        status:
          $ref: '#/components/schemas/HostStatus'
        total_cpus:
          type:
          - integer
          - 'null'
          format: int32
        total_memory_bytes:
          type:
          - integer
          - 'null'
          format: int64
...
    NewBootSource:
      type: object
      required:
      - name
      - kernel_image_id
      properties:
        description:
          type:
          - string
          - 'null'
        initrd_image_id:
          type:
          - string
          - 'null'
          format: uuid
        kernel_image_id:
          type: string
          format: uuid
        kernel_params:
          type:
          - string
          - 'null'
        name:
          type: string
    NewHost:
      type: object
      required:
      - name
      - address
      - port
      - host_user
      - password
      properties:
        address:
          type: string
        host_user:
          type: string
        name:
          type: string
        password:
          type: string
        port:
          type: integer
          format: int32
...
    NewStoragePool:
      type: object
      required:
      - name
      - pool_type
      properties:
        capacity_bytes:
          type:
          - integer
          - 'null'
          format: int64
        config: {}
        name:
          type: string
        pool_type:
          $ref: '#/components/schemas/StoragePoolType'
...
    StoragePoolType:
      type: string
      enum:
      - local
      - nfs
      - overlay_bd
...
    UpdateHostRequest:
      type: object
      required:
      - status
      properties:
        status:
          $ref: '#/components/schemas/HostStatus'
...